feat: update HomeService to include event retrieval and change ID type from Long to Integer in Department, Event, News, and User entities

+++ src/main/java/com/example/sdhzbozi/common/repositories/NewsRepository.java
@@ -3,5 +3,5 @@ package com.example.sdhzbozi.common.repositories;
 import com.example.sdhzbozi.common.model.News;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NewsRepository extends JpaRepository<News, Long> {
+public interface NewsRepository extends JpaRepository<News, Integer> {
 }
